+++ src/components/HeaderSection/Header/Header.tsx
@@ -3,9 +3,9 @@ import Box from '@mui/material/Box';
 import { useContext, useEffect, useState } from 'react';
 import Logo from 'src/assets/images/NavLogo.png';
 import ThemeSwitcher from 'src/components/HeaderSection/Switcher';
-import RegisterLoginModal from 'src/components/RegisterLoginModal/RegisterLoginModal';
 import LoginContext from 'src/contexts/LoginContext';
 
+import RegisterLoginModal from '../../SignInSignUpModal/RegisterLoginModal/RegisterLoginModal';
 import styles from './Header.module.scss';
 
 function Header() {
